Lab3-EC2ToS3/requirements.txt
requests==2.28.0
boto3
pandas
toml==0.10.2
python-dotenv==0.20.0